# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "edr-isobaric",
      "summary": "EDR API for isobaric data",
      "description": "See <https://api.met.no/weatherapi/isobaricgrib/1.0/documentation>",
      "termsOfService": "https://api.met.no/doc/TermsOfService",
      "contact": {
        "name": "The Norwegian Meteorological Institute",
        "url": "https://api.met.no/doc/support",
        "email": "weatherapi-adm@met.no"
      },
      "license": {
        "name": "Norwegian Licence for Open Government Data (NLOD) 2.0",
        "url": "https://api.met.no/doc/License"
      },
      "version": "0.9.0"
    },
    "paths": {
...
      "/collections/{collection_id}": {
        "get": {
          "tags": [
            "Collection Metadata"
          ],
          "summary": "Describe A Collection",
          "description": "Describe a specific collection.",
          "operationId": "describe_a_collection_collections__collection_id__get",
          "parameters": [
            {
              "name": "collection_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string",
                "pattern": "^isobaric$",
                "description": "Only available collection is isobaric",
                "title": "Collection Id"
              },
              "description": "Only available collection is isobaric",
              "examples": {
                "Isobaric": {
                  "summary": "The only available collection, isobaric",
                  "description": "Describe collection **isobaric**",
                  "value": "isobaric"
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Collection"
                      },
                      {
                        "$ref": "#/components/schemas/Collections"
                      }
                    ],
                    "title": "Response Describe A Collection Collections  Collection Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/collections/isobaric/position": {
        "get": {
          "tags": [
            "Collection Data"
          ],
          "summary": "Get Isobaric Page",
          "description": "Return data closest to a position.",
          "operationId": "get_isobaric_page_collections_isobaric_position_get",
          "parameters": [
            {
              "name": "coords",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string",
                "minLength": 9,
                "pattern": "^POINT\\(\\d+\\.?\\d* \\d+\\.?\\d*\\)$",
                "description": "Coordinates, formated as a WKT point: POINT(11.9384 60.1699)",
                "title": "Coords"
              },
              "description": "Coordinates, formated as a WKT point: POINT(11.9384 60.1699)",
              "examples": {
                "Oslo": {
                  "summary": "Henrik Mohns Plass 1",
                  "description": "Fetch data for a position in Oslo",
                  "value": "POINT(11.9384 60.1699)"
                },
                "Asker": {
                  "summary": "Asker",
                  "description": "Fetch data for a position in Asker",
                  "value": "POINT(10.45 59.83)"
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Coverage"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
...
        "Collection": {
          "properties": {
            "id": {
              "type": "string",
              "title": "Id"
            },
            "title": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "keywords": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keywords"
            },
            "links": {
              "items": {
                "$ref": "#/components/schemas/Link"
              },
              "type": "array",
              "title": "Links"
            },
            "extent": {
              "$ref": "#/components/schemas/Extent"
            },
            "data_queries": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DataQueries"
                },
                {
                  "type": "null"
                }
              ]
            },
            "crs": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Crs"
            },
            "output_formats": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Output Formats"
            },
            "parameter_names": {
              "$ref": "#/components/schemas/RootModel_Dict_str__Parameter__"
            },
            "distanceunits": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Distanceunits"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "id",
            "links",
            "extent",
            "parameter_names"
          ],
          "title": "Collection"
        },
        "Collections": {
          "properties": {
            "links": {
              "items": {
                "$ref": "#/components/schemas/Link"
              },
              "type": "array",
              "title": "Links"
            },
            "collections": {
              "items": {
                "$ref": "#/components/schemas/Collection"
              },
              "type": "array",
              "title": "Collections"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "links",
            "collections"
          ],
          "title": "Collections"
        },
...
        "Coverage": {
          "properties": {
            "id": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            },
            "type": {
              "const": "Coverage",
              "title": "Type",
              "default": "Coverage"
            },
            "domain": {
              "$ref": "#/components/schemas/Domain"
            },
            "parameters": {
              "anyOf": [
                {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/covjson_pydantic__parameter__Parameter"
                  },
                  "type": "object"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parameters"
            },
            "parameterGroups": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/ParameterGroup"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parametergroups"
            },
            "ranges": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/NdArray"
                  },
                  {
                    "$ref": "#/components/schemas/TiledNdArray"
                  },
                  {
                    "type": "string",
                    "minLength": 1,
                    "format": "uri"
                  }
                ]
              },
              "type": "object",
              "title": "Ranges"
            }
          },
          "additionalProperties": true,
          "type": "object",
          "required": [
            "domain",
            "ranges"
          ],
          "title": "Coverage"
        },
...
        "DataQueries": {
          "properties": {
            "position": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "radius": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "area": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "cube": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "trajectory": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "corridor": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "locations": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            },
            "instances": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/EDRQuery"
                },
                {
                  "type": "null"
                }
              ]
            }
          },
          "additionalProperties": false,
          "type": "object",
          "title": "DataQueries"
        },
...
        "Extent": {
          "properties": {
            "spatial": {
              "$ref": "#/components/schemas/Spatial"
            },
            "temporal": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Temporal"
                },
                {
                  "type": "null"
                }
              ]
            },
            "vertical": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Vertical"
                },
                {
                  "type": "null"
                }
              ]
            },
            "custom": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/Custom"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Custom"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "spatial"
          ],
          "title": "Extent"
        },
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "Link": {
          "properties": {
            "href": {
              "type": "string",
              "title": "Href"
            },
            "hreflang": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hreflang"
            },
            "rel": {
              "type": "string",
              "title": "Rel"
            },
            "type": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            },
            "title": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Title"
            },
            "length": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Length"
            },
            "templated": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Templated"
            },
            "variables": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Variables"
                },
                {
                  "$ref": "#/components/schemas/CubeVariables"
                },
                {
                  "$ref": "#/components/schemas/CorridorVariables"
                },
                {
                  "$ref": "#/components/schemas/ItemVariables"
                },
                {
                  "$ref": "#/components/schemas/RadiusVariables"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Variables"
            }
          },
          "additionalProperties": false,
          "type": "object",
          "required": [
            "href",
            "rel"
          ],
          "title": "Link"
        },
...
        "RootModel_Dict_str__Parameter__": {
          "additionalProperties": {
            "$ref": "#/components/schemas/edr_pydantic__parameter__Parameter"
          },
          "type": "object",
          "title": "RootModel[Dict[str, Parameter]]"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
...
      }
    },
...
  }